+++ project_folder/music_app_archive/tests/tests_models.py
@@ -12,6 +12,9 @@ User = get_user_model()
 class UserPlaylistsTest(TestCase):
     '''
     The fillowing test case contains positve and negative test cases for the Playlist model.
+
+    #TODO:
+        - Add playlist deleted 
     '''
     def setUp(self):
         self.user_1 = User.objects.create_user(