+++ project_folder/changelog.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+# 2025-02-08
+### Added
+* Selenium container
+
+### Removed
+* Redundant packages from environment.yml
+
+
 # 2025-12-09
 ### Added
 * Docker image