+++ project_folder/music_app_archive/changelog.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+
+# 2026-02-86
+### Fixed
+* Bandcamp bug via Selenium
+
+### Changed
+* Bandcamp API, we now scrape the HTML
+* Added mocking to Integrations API test in order to make the run time more efficient.
+
+### Added
+* Unit tests for the updated Bandcamp API
+
 # 2026-01-13
 ### Removed
 * Unsupported streaming platforms from the models and forms to avoid confusion